+++ tiago_lfc/launch/__init__.py
@@ -6,6 +6,8 @@ from .arguments import (
     all_arguments_from_yaml,
 )
 from .gz import (
+    GzWorld,
+    gz_control,
     gz_server,
     gz_spawn_entity,
 )
@@ -41,6 +43,8 @@ __all__ = [
     'all_arguments_from_yaml',
 
     # gz
+    'GzWorld',
+    'gz_control',
     'gz_server',
     'gz_spawn_entity',