+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2025-05-12
+
+### Added
+
+- `gz_control()`/`gz_control.launch.py`: New launch file use to control the sim
+  (start/pause/...) by cmd line
+
+### Changed
+
+- `gz_tiago_lfc.launch.py`: Now automatically starts the sim after spawning the
+  model
+
+### Fixed
+
+- `lfc_parameters.yaml`: Match with the new parameters of  LFC v2.0.0
+
 ## [1.0.2] - 2025-04-24
 
 ### Fixed
@@ -71,7 +87,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - README
 - pacakge.xml
 
-[unreleased]: https://github.com/ArthurVal/tiago_lfc/compare/v1.0.2...HEAD
+[unreleased]: https://github.com/ArthurVal/tiago_lfc/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/ArthurVal/tiago_lfc/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/ArthurVal/tiago_lfc/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/ArthurVal/tiago_lfc/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/ArthurVal/tiago_lfc/releases/tag/v1.0.0